api/requirements.txt
flask>=2.3,<3
requests>=2.31,<3
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
pandas>=2.1,<3
sqlglot==23.14.0
pg8000>=1.30,<2
